backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Cors==4.0.0
PyJWT==2.8.0
twilio==8.10.2
requests==2.31.0
python-dotenv==1.0.0
werkzeug==3.0.1
Flask-Admin==1.6.1
 